Allow loading const modules; Tests; Refactor
