+++ tests/modules/mod.rs
@@ -440,11 +440,34 @@ fn module_import_const_file() {
         sandbox.with_files(vec![FileWithContentToBeTrimmed(
             "spam.nu",
             r#"
-                    def foo [] { "foo" }
-                "#,
+                export def foo [] { "foo" }
+            "#,
         )]);
 
         let inp = &[r#"const file = 'spam.nu'"#, r#"use $file foo"#, r#"foo"#];
+
+        let actual = nu!(cwd: dirs.test(), pipeline(&inp.join("; ")));
+
+        assert_eq!(actual.out, "foo");
+    })
+}
+
+#[test]
+fn module_import_const_module_name() {
+    Playground::setup("module_import_const_file", |dirs, sandbox| {
+        sandbox.with_files(vec![FileWithContentToBeTrimmed(
+            "spam.nu",
+            r#"
+                export def foo [] { "foo" }
+            "#,
+        )]);
+
+        let inp = &[
+            r#"module spam { export def foo [] { "foo" } }"#,
+            r#"const mod = 'spam'"#,
+            r#"use $mod foo"#,
+            r#"foo"#,
+        ];
 
         let actual = nu!(cwd: dirs.test(), pipeline(&inp.join("; ")));
